Updated test folder structure and refactored tests
