+++ src/__tests__/App.test.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import { Provider } from 'react-redux';
+import store from '../redux/store';
+import App from '../App';
+
+test('renders App component without errors', () => {
+  render(
+    <Provider store={store}>
+      <App />
+    </Provider>,
+  );
+});
