+++ src/components/__tests___/Nav.test.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import { BrowserRouter as Router } from 'react-router-dom';
+import Nav from '../Nav'; // Replace with the correct path to your Nav component
+
+test('Nav component matches snapshot', () => {
+  const { asFragment } = render(
+    <Router>
+      <Nav />
+    </Router>,
+  );
+  expect(asFragment()).toMatchSnapshot();
+});
